ai-service/requirements.txt
fastapi==0.110.0
uvicorn[standard]==0.28.0
onnxruntime==1.17.1
numpy==1.26.4
Pillow==10.3.0
httpx==0.27.0
pydantic==2.7.0
python-dotenv==1.0.1
